requirements.txt
structlog
python-cloudflare
colorama
